@@ -4,9 +4,14 @@
   "private": true,
   "proxy": "http://server:8000",
   "dependencies": {
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.8.7",
     "axios": "^0.19.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.1",
     "react-scripts": "3.0.1",
     "webpack": "4.29.6"
   },
