@@ -1,1 +1,5 @@
-export { SampleComponent } from './sample-component/sample-component';
+export { ResponsiveAppBar } from './nav/nav';
+export { Visual } from './intro/visual';
+export { About } from './intro/about';
+export { ProjectCard } from './intro/project';
+export { Test } from './intro/test';
